Improve the time mechanic
Now the present is not deactivated when you travel in the past,
it is simply move to a background layer (invisible) and is still
simulated. This allow the past to corretly flow in the present even
when the player quickly goes from past to present (and vice-versa).
The past also override the position of the present object, so going in
the past "reset" the present.