fix past override mechanic
When the player goes in the past all the present states and positions
are override with the past ones, to avoid inconsistency this required
a copy method on every FSMState who keeps internal data that will copy
the past internal data.
A ghost listener issue was also fixed :
when copying the states the present statestack was flushed but the
exit() and Enter() methods were not called. Now they are.