Speed up time when going from past to present
Use timescale to briefly speed up time when going from past to present,
this allow to see only the cosequences of the past action when returning
in the present. To avoid spedding the player, his body is not simulated
during time transition (this will also allow for nice polish and
animation on the transitions and don't impact the feel).