Make player move with platform
A trigger makes the player a child of the platform so he moves with it.
However this require to be careful : when going from past to present the
past platfrom is deactivated so the player should not remain a child of
it, thus he is removes in the OnDisable() method.

Moreover, remove the transform update in the MoonlightSensitiveScripts
because it conflicts with other transform manipulation in the states (in
moving platform for instance) and create weird bugs. So know it is the
strirt responsability of the states to do any change on the object.