Create chain with distance joint
A chain is simply several gameobjects with a distance joint
component linked to the next gameobject. At the end a key can be
attached to the chain. Now the KeyHoldState has a chain field which can
be used to correctly handle the chain if the player hold the key : the
chain will be attached to the player as long as he kept the key, and
thus the player movement will be constrained.