Change the way to check if object are visible
Now the shadow renderer will also be responsible for chacking if objects
are visible, it do so by raycatsing toward the light for each vertex of
the collider, if nothing is hit then the object is visible.
A new boolean field "cast shadow" is available in moonLightSensitive
scripts and used by the shadow renderer to determine if the object
should or not cast shadow. If true, for each edge of the objetc collider
the shadow renderer will create a simple shadow polygon below, no more
raycasting involved.