Change core mechanic
There is no longer a past/present concept, from now on the core mechanic
is the superposition of two dimension :
- Objects of dimension 1 are visible in the shadow
- Objects of dimension 2 are visible in the light and cast shadows
This is achieved via stencil buffer and 4 shaders (light, shadow, past
and present).
The dimensions are still called "past" and "present" because it fits there
narrative and visual representation (and is still better than dimension
"1" and "2").

Now the player's interaction with one or another dimension based on the
light has to be implemented.